Ensure scroll bar appears on right side of screen
In some cases, when viewing a podcast’s episodes, the vertical scroll
bar which should appear on the right side of the screen to indicate how
far down the episode list you’ve scroll showed and animated in the
center of the screen.

To reproduce, revert this commit, search the podcasts for “J”, tap the
podcast named “Strip the Sound” (by some artist named “J”) and scroll
through the episodes. The odd behavior will be visible.

The solution used in this commit was found here:

https://github.com/facebook/react-native/issues/28085