Prevent infinite load when images come differently
When podcast images come in a different tags in the rss feed’s XML,
look for the image in all those places rather than loading indefinitely
when it’s not found in the only place we had been looking.

If the podcast artwork is not found anywhere, return undefined to avoid
crashing, continuing without podcast artwork.