Cache podcast episode lists
Some podcasts have hundreds or thousands of episodes.

When tapping the podcast in the search results list, you're transitioned
to a screen which loads the list of episodes. This is done with a GET
request to a webpage full of RSS text. The time it takes to load these
episodes and translate them from their XML format into javascript
objects can take several seconds.

As far as we know, there is no way around the initial load time being
long.

However, this commit enables the caching of episodes that you've viewed
before. Tapping to view a list of episodes which you've viewed before
shows the cached results instantly. A background get request fetches the
updated list and updates the displayed list (if necessary) when the fetch
completes.

Realm is the fastest on-mobile-device database according to our research.
We use realm for caching these episode lists.