Only resume after true audio interruptions
Podible was configured to pause when there were audio interruptions,
such as a turn by turn navigation announcements and resume playing at
the end of those audio interruptions. No matter what.

The problem with this is that if Podible is not playing, and a turn by
turn navigation occurs, after the turn by turn navigation announcement
ends, Podible resumes play even though the user hadn't been listening
when the interruption began.

In order to fix this problem, we need to make sure that when an audio
interruption ends, that the beginning of the audio interruption had
actually occurred while a podcast was playing.

There is a problem with this too:

There are native Objective-C hooks (see AudioInterruptions.m) which fire
off events triggering React Native / JS functions when audio
interruptions begin and end.

However, when an audio interruption begins, the audio is automatically
paused, so checking to see if the current playback state is "playing"
when the `audioInterruptionBegan` javascript function which is called
back into by the native code doesn't work (see that function in
`useAudioInterruptions.ts`).

Hence the need for this hacky usePrevious() hook, which manually keeps
track of the previous playback state's value.

With that, we can properly deduce whether or not playback should resume
at the end of an audio interruption, such as a turn by turn navigation
announcement.

This has been a big headache.