Display duration or time remaining for episodes
When viewing episode lists for a podcast, the total duration of each
episode is shown if the number of seconds listened to for that podcast
are 30 seconds or less.

When a podcast episode has been played completely, show “PLAYED” instead
of the total duration.

When a podcast episode has been listened for any amount in between the
above two mentioned timeframes, the time remaining is shown in the
format:

HH:MM:SS LEFT