Remove warning appearing after expo 40 SDK upgrade
The warning:
—

Constants.manifest is null because the embedded app.config could not be read. Ensure that you have installed the expo-constants build scripts if you need to read from Constants.manifest.

The solution:
—

add the following line to the Bundle React Native code and images Build Phase in your Xcode project:

../node_modules/expo-constants/scripts/get-app-config-ios.sh

The above excerpt was taken from this page:

https://github.com/expo/expo/tree/master/packages/expo-constants