Apologize when something breaks
We’re just learning how to effectively parse and update RSS feeds. It’s
possible for the feed parsing or updating to fail, which had been
leaving the user staring at an infinite loading spinner.

We know the particular spots where reacting to an RSS feed might break
things. If the rails api errors, then the server response with a 422
`unprocessable entity` status. When that happens, we apologize and state
that we’re fixing the problem.

When the app itself has trouble receiving the RSS feed, or something
goes wrong with the client side realm cache, we apologize and state that
we’re fixing the problem.