Fix bug causing first play after launch to break
Both of the two lines changed here are necessary to fix a bug.

Prior to this commit:

As a Podible user,
Given I have listened to an episode,
And I did not finish the episode,
When I quit and reopen the app,
And I see my “Recently Played” podcasts,
When I tap the podcast,
And I tap the unfinished episode (labeled with “hh:mm:ss LEFT”),
Then the podcast begins playing from the beginning.

This commit fixes the issue so that the podcast episode resumes playing
where it was left off.