Show updated time remaining on episode list screen
The episode list screen shows a time remaining label, eg if the podcast
is 10 minutes long and you've listened to 5 minutes, it displays:

00:05:00 LEFT

Prior to this commit, that label was only ever updated when the episode
list component first mounted.

After this commit, it is updated periodically, every second, so you can
watch it tick down.

This came at a slight cost to performance. JS frames / second sit at
about 58 rather than their ideal (and previous) 60 frames / second.

We understand that 55 frames / second and above is not noticeable by
the user, so we find this to be an acceptable tradeoff.

This commit also added ESLint for linting. I wanted it to help ensure I
was using hooks properly (in particular, making sure the dependency
lists for hooks like useEffect and useMemo were accurate).

Adding ESLint pointed out a number of errors elsewhere in the app, which
still exist in this commit. We'll address those errors in the next
commit, to keep the noise in this commit down, where the primary goal
was to update the time remaining label ticker periodically. The point of
this commit should remain clear and not be jumbled with all the
corrections ESLint is asking for.