Fix bug causing “crash” when Now Playing is viewed
We had been cancelling network requests to get episodes, if the user
navigates away from the episodes screen while the episode request was
in flight. Technically, viewing the “Now Playing” modal is navigating
away from the Episodes screen, and cancels the request. That behavior
is incorrect. We only want the request to be cancelled if the user
Navigates _out of_ the episodes screen, which is not the case when they
pull up the Now Playing modal over it.

This commit fixes that behavior so that the request to fetch episodes
isn’t cancelled when the Now Playing modal is pulled up.