Resume episodes where they’ve been left off
I hate this solution, but there are issues with beginning playing at a
seeked-to position in iOS for react-native-track player. This solution
seems to fix an issue causing the resume-where-you-last-left-off
feature to break and begin playing the podcast from the beginning.

The podcast where I could reproduce it was called Cortex. The setup to
reproduce it is to start playing a long episode (~2 hours or so), seek
to near the end of it so that about 5-10 minutes are left, then begin
listening to another episode and come back to the original episode.

Before this change, returning to the original episode did not resume
where the episode had been left off; it started it over.

After this change, returning to the original episode should resume
where you last left off.

I’m hoping to avoid ever needing to run this janky code in the future by
caching episodes and playing them from their cached download rather than
streaming them.