Collapse podcast descriptions
Podcast descriptions are sometimes super long, pushing the episode list
right off of the screen.

This commit adds a dependency which collapses text and shows a “more”
button at the end where the text has been truncated.

Tapping anywhere on the text (including the “more” button) expands the
description to be entirely visible, only pushing the episodes list down
and off of the screen after you already know they’re there / if and when
you want to.