Fix playback issue
Sometimes when resuming listening to an episode which has been started,
particularly when you’re resuming at a point long into the episode (say,
several hours into it), the resume just restarts the listener at 0:00.

I think this is because awaiting TrackPlayer.play doesn’t seem to work.

We avoid the need for awaiting TrackPlayer.play by using an event
listener which listens for a play event.