Prevent pulling up Now Playing screen from playing
There had been a bug where the useEffect we’d had in the Now Playing
screen which had intended to adjust the playback rate as it was slid
along the playback rate slider was causing playback to automatically
start if the audio was paused and someone tapped the up arrow on the
miniplayer to open the Now Playing screen.

This commit solves that issue by removing the useEffect in favor of
the use of a new hook, `usePrevious`.