Prevent scrub bar from flickering on play
When a podcast which has been listened to before is played, it takes a
while to fetch information about the audio file such as the duration
and reflect that in the position of the scrub bar. Instead of using
the reported position while the audio files is loaded / buffered, we use
the values which we know exist for duration and current position until
the correct times are loaded.

Otherwise, the scrub bar initially appears in the left most position
with elapsed and remaining time labels of “00:00” until the audio file
loads.