Fix bug causing resume play to buffer indefinitely
When listening to a track, if you pause the track, drop the Now Playing
Screen (dismiss it down off the screen), minimize the app, reopen the
app and then tap the play button on the MiniPlayer component at the
bottom of the screen, the play button turns into a loading spinner
indefinitely as the playback state switches to “Buffering” indefinitely.

This commit prevents the infinite buffering state from occurring and
allows resuming playback from the MiniPlayer in the way which you’d
expect.