Tap play to restart an already played episode
Prior to this commit, once an episode is finished playing, and the scrub
bar handle (the little dot) is fully right justified on the Now Playing
screen, tapping the play button causes the play button to infinitely
toggle between a play and pause icon very quickly. The episode does not
begin playing again.

After this commit, once an episode is finished playing, tapping the play
button restarts the episode from the beginning.